feat: search panel, hash path
